Move bootstrapper-related helper to common test utils project
