+++ test/Microsoft.Dnx.CommonTestUtils/TestUtils.cs
@@ -112,6 +112,33 @@ namespace Microsoft.Dnx.CommonTestUtils
             return process.ExitCode;
         }
 
+        public static int ExecBootstrapper(
+            string runtimeHomePath,
+            string arguments,
+            out string stdOut,
+            out string stdErr,
+            IDictionary<string, string> environment = null,
+            string workingDir = null)
+        {
+            string runtimeRoot;
+            if (string.Equals(Environment.GetEnvironmentVariable("DNX_DEV"), "1"))
+            {
+                // If DNX_DEV is set, then the path provided is to the root of a package.
+                runtimeRoot = runtimeHomePath;
+            }
+            else
+            {
+                runtimeRoot = Directory.EnumerateDirectories(Path.Combine(runtimeHomePath, "runtimes"), Constants.RuntimeNamePrefix + "*").First();
+            }
+            var program = Path.Combine(runtimeRoot, "bin", Constants.BootstrapperExeName);
+
+            string stdOutStr, stdErrStr;
+            var exitCode = Exec(program, arguments, out stdOutStr, out stdErrStr, environment, workingDir);
+            stdOut = stdOutStr;
+            stdErr = stdErrStr;
+            return exitCode;
+        }
+
         public static string GetMiscProjectsFolder()
         {
             var dnxRuntimeRoot = ProjectResolver.ResolveRootDirectory(Directory.GetCurrentDirectory());
