+++ test/Microsoft.Dnx.Tooling.FunctionalTests/project.json
@@ -1,6 +1,5 @@
 {
     "dependencies": {
-        "Bootstrapper.FunctionalTests": "1.0.0-*",
         "Microsoft.AspNet.Testing": "1.0.0-*",
         "Microsoft.Dnx.CommonTestUtils": "1.0.0-*",
         "Microsoft.Dnx.Runtime.Sources": "1.0.0-*",
